Upgrade navigator’s dependency on styled-components
This upgrades styled-components to the latest version and adds peer
dependencies as yarn told me to. I did test this a bit side-by-side
with Navigator from 1.2.0 to see if I could notice any changes both in
Firefox and Chrome and it looks exactly the same.

The changes are all fairly mechanical following type errors.

changelog_begin
changelog_end