Remove some dead code from transaction enrichment
When trying to understand how disclosure/divulgence information is
computed for the scneario service, I stumbled on some dead code:

* The `EnrichState.divulgeContracts` function is never used. Let's
  remove it.

* The `EnrichState.localDivulgendes` field is initialized with an empty
  map and never touched again. Thus, it's always going to stay an empty
  map. This means that the
  `EnrichedTransaction.localImplicitDisclosure` is also always going to
  be an empty map. Both fields can hence be removed.

I'm not sure if the `BlindingInfo.localDivulgence` field can be removed
as well. In my understanding, this data structure could be loaded from
transactions serialized with the field being non-empty in the past.
Thus, I've refrained from removing the field and set it to an empty map
when constructing a `BlindingInfo` from an `EnrichedTransaction`.

CHANGELOG_BEGIN
CHANGELOG_END