Feedback button (#6500)
* 'Thumbs up/down' and 'Feedback'  button fix

Changed the position of the 'Thumbs up/down' button to be in the bottom right corner above the 'Feedback' button. Added SASS to remove both buttons for mobile (as we were getting a lot of gibberish feedback

CHANGELOG_START
CHANGELOG_END

* 'Thumbs up/down' and 'Feedback'  button fix

Changed the position of the 'Thumbs up/down' button to be in the bottom right corner above the 'Feedback' button. Added SASS to remove both buttons for mobile (as we were getting a lot of gibberish feedback

CHANGELOG_START
CHANGELOG_END

* Putting some of the blocks where they were