Improve handling of over-application to allow much dead code to be eliminated. (#6586)
The dead code was hanging on by a slender thread. This change cuts that thread.

For over-apps, we now push a new continuation type KOverApp, instead of KArg.
KOverApp contains args of type SExprAtomic, instead of SExpr, and so its
execute() method may use enterApplication, instead of executeApplication.

This allowing lots of code to be removed:
- the continuation types: KArg, KFun, KBuiltin and KPap.
- defs: evaluateArguments and executeApplication.

changelog_begin
changelog_end