Improve the interface to execute (pure) builtins.
So that the resulting value is returned from the `execute` function, instead of being assigned to the `returnValue` machine state.

Rename the parent class from `SBuiltinMaybeHungry` to `SBuiltinEffect`, and have this be the parent of a few more subclasses which cannot meat the _pure_ interface.

Declare method `executeEffect` of the parent class `SBuiltinEffect` which has the same type as the original `execute`.

changelog_begin
changelog_end