Release a hotfix snapshot
Note for reviewers: This is a hotfix for a snapshot. I’m not going to
into why we need this here. As for the version number we decided that
the least confusing option here is the following:

Use the sha and number of commits as usual but use the date of the
original snapshot that we are patching here.

changelog_begin
changelog_end