Enter application with args already evaluated to values.
Combined with a small optimization for when `actualsSoFar` is empty, this means that we avoid constructing a new array for `actuals` and instead just use `newActuals` directly.

Also, in the case of an _over-application_, we can save just the over-applied `SValue`s (not args) in the `KOverApp` continuation, thus avoiding the need to save/restore the environment.

changelog_begin
changelog_end