Sequence the acquisition of connections in ledger-on-sql
Fixes #6716

The acquisition of a database connection, execution of a query and
further clean-up operations are now performed in a single callback to
make sure that a connection to the database is not attempted before the
previous query has been run, which could lead to unexpected failures as
a long query can prevent the acquisition of a connection for a time
longer than the connection acquisition timeout.

changelog_begin
[Sandbox] Fixed an issue that may have caused unexpected failures when
processing large transactions. Transaction can now be processed as long
as they fit in the queues configured using the back-pressure CLI
options. For more details:
https://github.com/digital-asset/daml/issues/6716.
changelog_end