remove click-to-zoom instruction
The ability to click on images to zoom was removed in #5233. This
removes the instruction to clic, as it doesn't do anything anymore.

Note: `git grep zoom docs` did not turn up any other instance, though
the lack of structure for the caption makes it hard to be sure.

CHANGELOG_BEGIN
CHANGELOG_END