daml-on-sql + sandbox + sandbox-classic: Split out custom CLI parsing.
I'm not sure this is the cleanest approach, but this is an attempt to
parse CLI arguments slightly differently depending on the version of
Sandbox we're using. For example, we can't disable implicit party
allocation in Sandbox Classic, so it shouldn't be an option, and Sandbox
Next doesn't support scenarios, so they shouldn't be around.

It also adds more tests for the CLI.