Mark pureconfig-macros as neverlink
pureconfig-macros pulls in the scala-compiler library which in turn
pulls in jansi-native which is somewhat problematic as it’s not
well supported outside of x86. It also just adds roughly 10% unused
junk to the http-json fat JAR. Afaict, pureconfig-macros is only used
at compile time so we can remove it.

changelog_begin
changelog_end