participant-integration-api: In `JdbcIndexer`, log with context. (#7042)
* participant-integration-api: In `JdbcIndexer`, log with context.

We were not providing the correct `loggingContext` to
`JdbcIndexer#handleStateUpdate`. This means we were just dropping useful
information. This adds the implicit so that it uses the correct logging
context.

There's a bigger problem, in that there are multiple logging contexts in
scope, making this very error prone. We'll need to figure out a way to
avoid this as much as possible.

CHANGELOG_BEGIN
CHANGELOG_END

* participant-integration-api: Purge unnecessary newlines in JdbcIndexer.